Break out modules for Verilog reading and the Design class. Fix Spice SUBCKT case-sensitivity. Fail when encountering Verilog ast.Assign
